requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# LLM
openai>=1.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg
psycopg2-binary
alembic

# Validation & Config
pydantic>=2.0
pydantic-settings

# Privacy
presidio-analyzer
presidio-anonymizer
spacy>=3.0

# Reports & Charts
matplotlib

# Scheduler
apscheduler>=3.10

# Environment
python-dotenv
